Moved CPU relevant implementation from external/hardware to external/cpu Implemented rescaling of energy and power according to cpu utilization
